requirements.txt
streamlit
pandas
plotly
numpy
chromadb
google
google-genai
google-api-core
python-dotenv
scikit-learn
matplotlib